[FIX] website_sale(_stock): move test due to dependency to stock
In odoo@98ca5140baf96, a test using a product of type 'product' was
added. Since this type is introduced in stock, stock needs be installed.
Since the delivery module was split in saas-16.2 to remove the
dependency to stock, this test needs to be moved in website_sale_stock.

Build error 48757

X-original-commit: 3e28d9b2b47245404e9341082f5663174ad0df16
