+++ addons/website_sale_stock/tests/__init__.py
@@ -4,5 +4,6 @@
 from . import test_website_sale_stock_abandoned_cart_email
 from . import test_website_sale_stock_multilang
 from . import test_website_sale_stock_product_warehouse
+from . import test_website_sale_stock_delivery
 from . import test_website_sale_stock_stock_notification
 from . import test_website_sale_stock_reorder_from_portal
